yolo_backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
ultralytics==8.3.40
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pillow==11.0.0
arabic-reshaper==3.0.0
python-bidi==0.6.3
deep-translator==1.11.4